+++ src/schemas/goals.py
@@ -1,0 +1,10 @@
+from pydantic import (
+    BaseModel,
+)
+
+
+class Goal(BaseModel):
+    period: str
+    type: str
+    target: int
+    unit: str
