+++ src/schemas/requests.py
@@ -1,0 +1,17 @@
+from typing import Optional
+from pydantic import (
+    BaseModel,
+    EmailStr,
+)
+
+
+class UserCreate(BaseModel):
+    email: EmailStr
+    password: str
+    full_name: Optional[str] = None
+
+
+class AdminCreate(BaseModel):
+    email: EmailStr
+    password: str
+    full_name: Optional[str] = None
